feat: allow to pass a config flag on cli and server cmds
By default, the cli tools are picking the config from config.yaml
on the same directory. While it is the default, it is desirable
to pass a config flag that will allow to manually specify the
config file path.

Closes: #112
